ADD implementato metodo checkDatePrenotazione dentro a PrenotazioniService ed eseguito il controllo su update e delete delle prenotazioni per non far modificare dati meno di 2 giorni prima della partenza

+++ src/main/java/com/rentalcar/webapp/service/PrenotazioniServiceImpl.java
@@ -6,6 +6,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.Date;
 import java.util.List;
 
 @Service("PrenotazioniService")
@@ -27,16 +30,29 @@ public class PrenotazioniServiceImpl implements PrenotazioniService {
 
     @Override
     public void update(Prenotazioni prenotazione) {
-        prenotazioniDao.update(prenotazione);
+        if (checkDatePrenotazione(prenotazione.getStartdate())){
+            prenotazioniDao.update(prenotazione);
+        }
     }
 
     @Override
     public void delete(Long id) {
-        prenotazioniDao.delete(id);
+        if(checkDatePrenotazione(this.getPrenotazione(id).getStartdate())){
+            prenotazioniDao.delete(id);
+        }
     }
 
     @Override
     public Prenotazioni getPrenotazione(Long id) {
         return prenotazioniDao.getPrenotazione(id);
     }
+
+    @Override
+    public boolean checkDatePrenotazione(Date start) {
+        LocalDateTime ldt = LocalDateTime.ofInstant(start.toInstant(), ZoneId.systemDefault()).minusDays(2);
+        if (ldt.isAfter(LocalDateTime.now())){ return true;}
+        else {return false;}
+    }
+
+
 }
